PEP8 fixups
There were a couple of pep8 violations that are fixed in this patch. To test:

`pep8 sat6Inventory.py`

This patch does not address `E501` line too long errors but should take care of
the others. Note that I only did cursory testing on this, and did not test
against a live server.