try to fetch list of hosts w/o details
In http://projects.theforeman.org/issues/20072 a new param was added:
 thin
It allows to fetch a list of hosts w/o details, which makes fetching a
lot faster. And we need to fetch the full details later anyways.

Poor mans benchmark with about 100 hosts and patched Satellite:
sat6inventory runs 3:00min without the patch, 2:40min with the patch